리팩터: 오류 핸들러들을 미들웨어로 옮김
